requirements.txt
Django==5.0.4
django-filter==24.2
djangorestframework==3.15.1
drf-yasg==1.21.7
pillow==10.3.0
psycopg2-binary==2.9.9
pytest==8.2.1
pytest-django==4.8.0
python-dotenv==1.0.1
